Editions can be marked as being in beta
www.agileplannerapp.com/boards/173808/cards/5418

As a GDS content designer
I want to show GOV.UK end users that a piece of
mainstream content is not yet the official source
and is still in beta
So that I can trial content with the public and
gather evidence to make it even better
