+++ app/assets/stylesheets/application.scss
@@ -1,5 +1,7 @@
 @import 'govuk_admin_template';
+@import 'colors';
 @import 'bootstrap_and_overrides';
+@import 'forms';
 
 // Components
 @import 'sidebar_nav';
