+++ app/assets/stylesheets/colors.scss
@@ -1,0 +1,1 @@
+$beta-color: #f47738; // orange
