+++ app/assets/stylesheets/forms.scss
@@ -1,0 +1,5 @@
+#edition_in_beta_input {
+  .help-block {
+    margin-top: 0;
+  }
+}
