+++ app/views/root/_publication.html.erb
@@ -21,6 +21,10 @@
     <% if publication.has_video? %>
       <i class="glyphicon glyphicon-film" alt="Has video"></i>
     <% end %>
+
+    <% if publication.in_beta? %>
+      <span class="badge badge-beta">beta</span>
+    <% end %>
   </td>
   <td>
     <%= publication.slug %>
